refactor (src): change determinant into fallen and change unit into measurement
