+++ include/kansei/node/inertial_manager.hpp
@@ -31,8 +31,8 @@
 #include <string>
 #include <vector>
 
-#include "kansei/determinant/determinant.hpp"
-#include "kansei/unit/unit.hpp"
+#include "kansei/fallen/fallen.hpp"
+#include "kansei/measurement/measurement.hpp"
 
 namespace kansei
 {
